+++ packaging/conan/conanfile.py
@@ -8,7 +8,7 @@ class HeliosNetworkConan(ConanFile):
             self.options.rm_safe("fPIC")
 
     name = "helios-network"
-    version = "0.10.2"
+    version = "0.10.3"
     license = "MIT"
     url = "https://github.com/helios-graphs/helios-network"
     description = "Helios Network native graph core compiled as a C library"